Change import path to reflect GitHub location
The code was building fine in my local layout, even when writing other
code that tried to access the example package `egpack`. However, anyone
else trying to use that package would need to use a different path,
based on the GitHub location. So change my code to use the same form.